requirements.txt
python-dotenv
loguru
requests
beautifulsoup4
playwright
feedparser
praw
httpx
ollama
qdrant-client
groq
google-generativeai
fastapi
uvicorn
streamlit
apscheduler
langdetect
Pillow
pandas
numpy
pydantic